LPS-88181 Enforce a long cast to prevent SQL injection.

+++ portal-kernel/src/com/liferay/portal/kernel/service/persistence/impl/BasePersistenceImpl.java
@@ -260,7 +260,7 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 		}
 		else {
 			for (Serializable primaryKey : uncachedPrimaryKeys) {
-				query.append(primaryKey);
+				query.append((long)primaryKey);
 
 				query.append(",");
 			}
